Resurrect Larry Kollar's ms manual.
Per recent discussion on the groff development list[1], bring back the
ms version of the ms manual.  Expansions and corrections from the
corresponding chapter of our Texinfo manual are to be ported to this
document, and then the Texinfo chapter deleted.

Thanks to Larry Kollar for locating (and writing!) this document.

[1] https://lists.gnu.org/archive/html/groff/2020-06/msg00044.html
    et seq.