Adapt to the fact that \: reënables hyphenation.
...and rename "hyphenless [discretionary] break" to "non-printing break
point" in documentation.

In spite of my discomfort at learning this fact, I decided not to try to
change groff's behavior because (1) frankly, doing so looked challenging
in the implementation and (2) it occurred to me that, for those
circumstances in which the present behavior is desired by the user, the
mechanism of recovering it would be _even more_ challenging.  Keeping in
mind that requests and escapes are "assembly language for typesetting",
I'm leaving the language alone and adapting to it.  The pattern '\:\%'
will be unambiguous to locate and easy to change later if someone has a
brighter idea.

* doc/groff.texi (Manipulating Hyphenation): Rename "hyphenless break"
  to "non-printing break point".  Clarify that it is the soft hyphen
  glyph, not necessarily a hyphen, that is ordinarily written to the
  output on hyphenation breaks.  Note that the remainder of a word after
  \: is subject to hyphenation as normal.  Note (new) '\:\%' idiom for
  getting what people will want at least some of the time.  Update
  example to use it.

* makevarescape.sed: Insert \% after we insert \: to rewrite slashes in
  filenames, to protect later portions of the filename from hyphenation.

* man/groff.7.man (Escape Sequences/Escape short reference): Rename
  "hyphenless break" to "non-printing break point".

* man/groff_diff.7.man (Language/Escape sequences): Resync with our
  Texinfo manual.

* tmac/groff_man.7.man.in (Description/Hyperlink and email macros):
  Rename "hyphenless break" to "non-printing break point".  (The
  examples did not need updating because the macros already disable
  hyphenation--too aggressively as it turns out, but that's a different
  change set...)