gperl(1): Get rid of page-local bullet string.
This is one of the more superfluous non-portabilities I've seen in the
groff man page corpus.  There's no need to define a string for a
typographical bullet.  The bullet special character escape \(bu is
documented in CSTR #54.  Possibly Bernd didn't like what was substituted
for it on some nroff devices, but -Tutf8 _does_ have a proper bullet
character (depending on your terminal's font), and if a user of a less
capable terminal doesn't like the fallbacks in tty.tmac (I'm not crazy
about the simple use of "o" with -Tascii on glass TTYs myself), they can
do this in their man.local:

	.if '\*[.T]'ascii' .char \[bu] *

...which achieves what Bernd was doing, but for all man pages.

Testing did reveal something interesting, though...

I had to go ahead and delete the "2m" indentation argument to .IP as
well.  Without it, the body of the paragraph got pushed to the next
output line, as if the tag were too wide.  Obviously neither a "*" nor
a bullet is wider than 2 ems.  Experimentation reveals that an indent of
"2n" leads to a break, but "3n" does not[1].  It seems there is an
off-by-one error either in our documentation of .IP and .TP, or in the
implementation.  If this was the real reason for the string definition,
it was undocumented and obscure.  My Gordian solution was to dispose of
the indent argument to .IP calls as well.

[1] 1m - 1n = 0 on nroff devices.