tmac/an-ext.tmac: Succumb to hyphenation reality.
URLs in real documents, like our man pages, were getting hyphenated
in spite of my efforts.  Fix this by permitting the "punctuation"
arguments to .ME and .UE to be hyphenated again.

Here's my understanding.

I think this is because while it is tempting to think of a groff
document as a tree (and if we were redesigning it from scratch, it might
be), that is not completely true.  Changes in hyphenation mode don't
leave nodes in the tree that will be respected when it's time to break
the line.  Instead, at break time, groff checks the current hyphenation
mode, and if it's not 0, merrily backtracks over the pending output line
and sticks in a hyphen wherever the patterns and mode will allow and
where not specifically prohibited by hyphenation exceptions or words
beginning with \%.

* tmac/an-ext.tmac (.ME, .UE): Restore hyphenation after "punctuation"
  arguments have been output.  The next token will always be a space
  node, the end of the document, or similar.

* tmac/tmac.am: Mark the punctuation hyphenation tests as expected to
  fail, because they now do.