groff_mom(7): Drop pdfmom from synopsis.
There's no reason to have pdfmom in the synopsis; this clashes with the
man pages for all other macro packages.  Admittedly, most of them don't
have wrapper commands, but the one that does, mmroff(1), also doesn't do
this.  Furthermore, pdfmom's man page is more diligently maintained and
Peter Schaffter is dubious about the utility of this one.  pdfmom is
referenced in the Description section.

Also add reference to section "See Also".