man pages: Regularize editor aid comments.
* Add fill-column (Emacs) and textwidth (Vim) where missing.
* Sort Emacs Local Variables alphabetically as is done elsewhere.
* Drop junk comments.

Not every man page has an identical comment block of this type.  (1) A
couple of man pages use "nroff" instead of "groff" for the Vim filetype
because they use no groff extensions at all.  (2) A couple of man pages
are encoded in Latin-1 instead of plain ASCII.  (3) A few man pages use
a custom tab-width (Emacs) and tabstop (Vim) to ease table maintenance.