Document groff_man(7) adjustment control.
Ordinarily I'd commit the documentation with the code change as is
proper, but I broke them up in this case so I could test the change to
the page in light of the one to an-old.tmac.  With "groff -man -dAD=l"
now recognized, my diffing procedure enables me to review the man page
changes without all 60 pages showing me the difference between ".ad b"
and ".ad l".

* tmac/groff_man.7.man.in (Strings) <\*(AD>: Document it.

  (Strings) <\*(HF>: Parallelize language with new \*(AD.

  (Options) Expand introductory sentence to discuss -d and -r options.

  (Options) <-dAD>: Document default and likely values.  Nobody
  right-justifies or centers man page text; send the curious to
  groff(7) for '.ad' request documentation.

  (Options) <-dHF>: Document default and steer people to groff(7) for
  '.ft' request documentation.