groff_man*(7): Update history of \*[Tm].
* tmac/groff_man.7.man.in (Description/History): The Tm string is eight
  years older than I thought.  It seems to have come from PWB/UNIX 2.0,
  or maybe a field upgrade thereof.

(Pleasantly, this allows me to efface my embarassing mistake of writing
"Ultrix 11" as if "11" were a version number.  I should have known
better given that I knew about the MACRO-11 assembler and the popularity
circa 1980 of suffixing software products with a nod to the supported
ISA ("-80" for 8080 and Z80 machines, "-65" for the 6502, "-68" for
m68k, and so on).  The first DEC machine I ever actually saw in person
was an Alpha.  Yup, I'm a whipper-snapper.)