doc/groff.texi (Setting Registers): Add reference.
...to Auto-increment node.  Thanks to Marc Chantreux for the suggestion.

Also screen the node for quote usage and bracket it with @codequote*
commands to the syntax summaries and examples don't look ugly.

To that end, wrap the apostrophes in the syntax summary of the
(manually) incrementing \R escape in @code{}.

Also say "GNU troff" instead of "gtroff" or plain "troff".