groff_char(7): Replace Latin-1 table.
The purpose of this table originally, I suppose, was to show Latin-1
groff users all the cool glyphs that were supported as input by the
engine.  Today, Latin-1 is old and busted and UTF-8 is the new hotness.
(Okay, that was just as true twenty years ago.)

But there were other problems.

The old one was too wide, far overrunning the 78n limit in nroff mode.
It featured a column, "Code", that the subsequent glyph tables lacked,
and which was blatantly duplicative of the decimal \[charNNN] escape
notation.  Said escape notation was not itself escaped, rendering the
"Input" and "Output" columns the same, making the table far less useful
for users who didn't know how to compose all of the glyphs in the table
with their keyboards.  (Okay, it's wasn't difficult to guess that one
could use the "Code" column to construct a \[charNNN] escape, but that
observation renders the "Input" column superfluous for a different
reason.)  Also, the table had one row per code point with no blank
lines--the longest such stretch by far in this man page--which
frustrated tbl(1) and made it complain about not being able to fit it on
a page.  (To be fair, that tbl(1) complains even in the
continuous-rendering mode used by man pages is probably a bug in it.)

The new table uses only three columns; the rendered glyph, the special
character escape that renders it, and a glyph description.  This turned
out to narrow the table enough that I could arrange in into two adjacent
columns of 3 fields each (with literally NO room left for horizontal
expansion--it's right at the 78n limit, thanks to "{femin,mascu}line
ordinal indicator").

What's more, the input column now documents glyph names instead of the
legacy eight-bit numeric special character escape form.  These are
shorter, more mnemonic, and portable to EBCDIC systems.  Some of them
are even portable to AT&T troff and its descendants.

Since every single one of these glyphs appears in a later table, no
information is lost by dropping the AGL and Unicode code point columns.