NEWS, ChangeLog: Clarify the history of \sN.
The weirdness of legacy \sN parsing was in fact documented some years
before 1992, but unfortunately not in a place where users might think to
look for it: background material in CSTR #97, the paper describing
Kernighan's device-independent refactor of AT&T troff.

In this document, dated as "Revised, March 1982", Kernighan writes:

"TROFF turns out to be surprisingly dependent on the Graphic Systems CAT,
not just in the code but in its design.

Some of the design dependencies are pretty obvious.  For example, the
CAT provides four fonts (of 102 characters each) and 15 sizes.  The
specific sizes are wired into the syntax of the language: since the CAT
has no sizes larger than 36 points, \s46 can be uniquely decoded as a
4-point 6, while \s36 is simply a switch into size 36."

It will require grayer beards than mine to opine on what proportion of
troff users had on-demand access to CSTR #97 in the 1980s; being of
1982 vintage, I surmise that this paper came too late to be included in
the commercially-sold volumes of Version 7 Research Unix manuals.