makevarescape.sed: Use for \[] special characters.
The only man page we have that doesn't enter compatibility mode (neqn)
also doesn't use any sed-substituted patters where characters replaced
by a \[] special character escape form are likely to be used.

The @g@ command prefix is the main avenue for intrusion, but I think it
improbable that many people are going to include apostrophes, double
quotes, carets, grave accents, or tildes in the command prefix; some or
all of these will pick fights with the shell and require quoting that we
don't represent in man page text anyway.

On top of that, neqn is largely a stub page.

If it's a problem, the better fix is to simply make neqn switch out of
compatibility mode like all our other man pages.