doc/groff.texi, groff_diff(7): Tweak example.
It's like bathing a cat to get programmers to punctuate inset examples
that are grammatically part of a sentence, as if they were part of a
sentence.  People love to just trail off with "the following:", blitz
the user with 1-100 lines of block-quoted literal, and then pretend they
ended a sentence.  (This is suggestive of the console jockey ethos; once
one has thrown code at you, they're done with you.)

Mathematical literature is more careful.

However, the man macros make doing the right thing difficult; in nroff
mode there just aren't many font styles to go around, and while one
could ".if t" around that problem, a miniature BSD daemon materializes
on my shoulder and pokes me with a pitchfork if I even consider it.  And
doing so doesn't help the likely majority of people who _only_ read man
pages in a terminal window.

In the end the best solution is to just recast.  Our man pages are
already overrun with lazy "sentences" ending in "the following:" and
similar constructions anyway, so it's no burden to scrub out one more.

tl;dr: A colon does not relieve you of your duty to finish a sentence.