tmac/mdoc/doc-common-u: Fix style problems.
Correct misleading indentation.

Also fix spelling of register name.  Despite appearances, this wasn't
actually a typo but a deliberate mispelling; I learned in the course of
implementing 0438b1b905ebe9ac5fc678af06db911d25c3a030 (14 September)
that removing a non-existent register was not an error (or even a
warning) condition.  I looked briefly into adding code to warn about it,
found it difficult--the string that's used for lookup in the register
dictionary seems to not actually be in the scope of `remove_reg()`,
which implements the '.rr' request--at any rate it's not a "symbol"
because undefined symbols are reasonably enough not in the troff symbol
table, and I don't know the class hierarchy well enough to guess where
else it might be.  I then forgot to revert the change I was using to try
to provoke the diagnostic, and didn't notice because...there's no
diagnostic.  (One could still be emitted from `remove_reg()`, but it
wouldn't be able to include the name of the unrecognized identifier, and
I consider that too user-hostile to bother with.)  Anyway, this is an
object lesson in using "XXX" comments.