groff_man*(7): Reduce redundancy.
Undocument \*(AD and \*(HY in subsection "Strings", since they're
already documented in section "Options".  Also, they're not part of the
page authors' interface to the macro package, but rather the page
_readers'_ interface.  Add internal cross-reference.

Also this makes the page(s) shorter.

Recast description of \n(SN to be less jarring; there's no reason to
suggest that a page has only one subsection heading (usually there are
zero or several).