tmac/an-old.tmac: Clarify \n[C] setup for HTML.
Make interaction of \n[C] register and HTML output more clear in code
and comments.  It looked strange to test \n[C] for "falseness", throw a
warning about that not being supported for HTML, and then force its
value to false anyway.  This was harmless given the following logic, but
confusing and possibly not robust against future refactoring.  Do the
redundant setting and explain why in a comment.