groff_char(7): Fit Latin-1 table into 78 columns.
...for real this time.  It was still overrunning on the "ascii" output
device, thanks to the substitutions of the 3-cell "(C)" and "(R)" for
© and ®, respectively.

Drop "congruent to" from notation of \[=~]; it just won't fit, and I
don't want to go back to using abbreviations in the descriptions.

Remove XXX comments about tables overrunning 78n, because, thanks to the
elimination of the AGL column in the previous commit to this file and
the change above, they no longer do.