groff_filenames(5): Remove bad information.
The claim about Heirloom Doctools troff using "1b" as a man page
extension appears to be false; as far as I can tell, this is a suffix
used only in its source tree (as illustrated by SCCS ident strings that
persist to this day).  But in that case, they have no more meaning than
an "*.in" suffix, and in event they aren't present nowadays.

The claim about andoc "files" having a dedication extension
is...vacuous.  You can't write a file "in" andoc--the whole point of
andoc is to sense whether mandoc(7) or man(7) is the language of the
source document.