[ms test]: Fix typo and rename file.
One character slipped away as part of a regex rewrite.  The test works
and remains working, but this way it will also detect a page numbering
error in the main matter.  Which is an abuse of scope, but this is the
only ms test we _have_.  Though that's about to change.

Also rename test file per the convention in use.