groff_mdoc(7): Delete misleading sentence.
Ingo helpfully followed up with regard to my XXX comment.  I had
misunderstood the existing page language to suggest that mdoc(7) does
something especially clever in this area.  His surmise is that it is
there to help the reader understand that literal double quotes don't
survive macro calls.

It is not quite the truth to tell people that, but it's close.  You can
get double-quotes into macro arguments without .tr tricks or the groff
\(dq special character escape, but I sure as hell wouldn't want to try
to explain how.  These are the arcane quoting rules I've mentioned in
some of my other documentation commits.

It's been a while since we've played Predict *roff Behavior.  Ready?

.\" Let's count double quotes in binary.
.nr n -1 +1
.af n 00
.de M
\\n+n \\$1 \\$2
.br
..
.M  a    b \"
.M  a    b"\"
.M  a   "b \"
.M  a   "b"\"
.M  a "  b \"
.M  a "  b"\"
.M  a " "b \"
.M  a " "b"\"
.M  a"   b \"
.M  a"   b"\"
.M  a"  "b \"
.M  a"  "b"\"
.M  a""  b \"
.M  a""  b"\"
.M  a"" "b \"
.M  a"" "b"\"
.M "a    b \"
.M "a    b"\"
.M "a   "b \"
.M "a   "b"\"
.M "a "  b \"
.M "a "  b"\"
.M "a " "b \"
.M "a " "b"\"
.M "a"   b \"
.M "a"   b"\"
.M "a"  "b \"
.M "a"  "b"\"
.M "a""  b \"
.M "a""  b"\"
.M "a"" "b \"
.M "a"" "b"\"

Do any double quotes appear in the output?  If so, name the integers
that have them, and where.

It may be worth noting that groff HEAD (-Tutf8) and Heirloom doctools
nroff, release 191015, produce identical output for the above.