tmac/an-old.tmac: Add style checks.
Add register, CHECKSTYLE, to capture user preference for emission of
style-related diagnostics.  Defaults off.

Not documented yet because I expect there to be some discussion on the
mailing list over whether it should be and/or what the register should
be called.

Whatever emerges from that discussion can and should be parallelized in
support and form with diagnostics in our mdoc(7) implementation, which
has some style-oriented diagnostics that issue at the same level as more
serious problems.

* tmac/an-old.tmac: Initialize CHECKSTYLE to false if not already set.
  (an-style-warn): New macro emits diagnostic of type "style" when
  called if CHECKSTYLE register is true.
  (TH, RI, IR, IB, BI, RB, BR): Call an-style-warn if fewer than two
  arguments are seen.