groff(7): Fix rendering nits.
The previous commit to this file caused spurious styling to be applied
to punctuation characters in a few cases.  Rather than further
engineering the page-local macros to the level of robustness required to
handle those instances, simply convert them to use standard man(7)
macros.  As a side effect, this eliminates several uses of the \c
escape, which frightens people.

Also fix an outright rendering error of much older vintage (2009) caused
by the pounding of square \O0 and \O1 pegs into the round hole of the
"clever" .escarg page-local macro, which assumed that its second
argument would be a non-literal parameter and set it in italics.  This
is of course incorrect for literal uses of "0" and "1".  In these cases,
laissez .escarg go.