[docs]: Document groff/troff -a option.
As far as I can tell, I'm the first person ever to document what this
option does in anything but the vaguest terms.  :-/  I checked DWB 3.3
troff, Solaris troff, Plan 9 troff, Heirloom Doctools troff, and
neatroff (the last of which doesn't implement it).

Thanks to Dave Kemper for the discussion in Savannah #55278.