Lower new unplanted trap error to 'mac' warning.
* src/roff/troff/div.cpp (top_level_diversion::change_trap): Change
  error upon invalid attempt to move an unplanted trap into a warning of
  type 'mac'.

This somewhat stretches the documented semantics of the 'mac' warning
type, so recast them to accommodate the new instance.

* doc/groff.texi (Warnings):
* src/roff/troff/troff.1.man (Warnings): Recast description of 'mac'
  warning type to include the above scenario.

  Recast to mention boxes as well as traditional diversions.  Refer to
  macros, strings, diversions, and boxes collectively as "objects"
  instead of "macros", avoiding synecdoche.  Use complete sentences.

Thanks to Bjarni Ingi Gislason for reporting the diagnostic arising in
real life, and to Peter Schaffter for the discussion and recommendation.
Some bike sheds get painted quickly!

Fixes <https://savannah.gnu.org/bugs/?59573>.