Document use of SOURCE_DATE_EPOCH and TZ.
The semantics of SOURCE_DATE_EPOCH to groff were not established with
respect to time zone selection, prompting divergent interpretations;
Debian and distributions derived from it have for several years patched
groff to implicitly use UTC as the time zone when interpreting the
current time (or SOURCE_DATE_EPOCH) as a local time.  While a convenient
and defensible choice for reproducible build efforts[1], it runs against
the grain of user expectations.  Systems programmers like monotonically
increasing clocks; the broader user base usually prefers a clock that
follows an applicable civil calendar.  To the latter audience, a
difference between
  $ date "+%Y-%M-%d %H:%m:%S"
and
  $ groff <<EOF
  .af year 0000
  .af mo 00
  .af dy 00
  .af hours 00
  .af minutes 00
  .af seconds 00
  .tm \n[year]-\n[mo]-\n[dy] \n[hours]:\n[minutes]:\n[seconds]
  EOF
is difficult to explain.

* doc/groff.texi (Environment):
* src/devices/grohtml/grohtml.1.man (Environment):
* src/devices/gropdf/gropdf.1.man (Environment):
* src/devices/grops/grops.1.man (Environment):
* src/roff/groff/groff.1.man (Environment):
* src/roff/troff/troff.1.man (Environment): Expand description of use of
  SOURCE_DATE_EPOCH to note that it converts the time (as seconds since
  the Epoch) to local time, which motivates documentation of the TZ
  environment variable.  Explain what the time stamps are used for:
  troff merely stores the human-readable time components in registers;
  the grohtml, gropdf, and grops output drivers write the information to
  the output in comments.

Fixes <https://savannah.gnu.org/bugs/?57218>.

[1] https://reproducible-builds.org/docs/source-date-epoch/