doc/me{intro,ref}.me: Use consistent terminology.
Historical and modern *roff documentation is careful to exclusively use
the term "request" for what coders more conventionally call a "command."

doc/meref.me largely follows this practice, but slips up in a couple
places, using "command" for *roff requests.  This patch fixes those.

While it's at it, it fixes a few places where "request" or "command"
erroneously refers to a macro.

The situation in doc/meintro.me is trickier.  This document pervasively
calls macros defined by the -me package "requests," and reserves the
term "macro" for macros the user defines in his own -me document.  While
not strictly accurate, this slight muddying is consistent and probably
fine for the intended (novice) audience of this document.

The sole exception to this consistency is text I wrote that was added in
commit 7a83f7f50, which calls .bl a macro--technically accurate but
inconsistent with the rest of the document's usage.  Thus this patch
changes this word to "request."

The one use of "command" in meintro.me is also changed to "request"
following similar logic.