groff_char(7): Add plain characters to tables.
In the handful of cases where a regular input character is a synonym for
a special character escape, add it to the glyph tables.  This was
already done in part.

Add paragraph to subsection "Quotation marks" emphasizing the neutral
double quote's domain of application as being areas like programming
languages, implying its inadequacy for traditional typographical
purposes.