[troff]: Ignore out-of-range .ad argument.
* src/roff/troff/env.cpp: Add ADJUST_MAX enumeration constant to record
  maximum valid numerical adjustment mode.
  (adjust): Verify numeric argument against ADJUST_MAX instead of
  numeric literal.  Ignore excessively large values instead of setting
  adjustment mode to "right".  Update warning diagnostic.

This is a behavior change, albeit to how an invalid argument is
interpreted.  What we call "AT&T troff" in our documentation is a
reference to CSTR #54 in its 1992 revision, which we use as a
specification because source code to the implementations was not freely
available.  Because the above circumstance is out-of-spec, "AT&T troff"
says nothing about it and offers no precedent.

Given that, let's have a look at what other *roffs do.

* V7 Unix troff ignored invalid adjustment modes of all kinds, falling
  off the end of a switch with no default case.
* DWB, Plan 9, and Solaris troff make no changes to V7 Unix troff here.
* Heirloom Doctools troff and Neatroff extend the range of valid numeric
  arguments to support paragraph-at-a-time formatting.
* Neatroff furthermore supports an '.ad k' mode for support of
  keshideh/kadisha justification of cursive scripts.

GNU troff's historical behavior here does not seem useful, and prepares
users poorly for adjustment modes groff may support in the future.