tmac/tmac.am: Remove unidiomatic for loop.
* tmac/tmac.am ($(TMACSTRIPFILES)):
  ($(TMACMDOCSTRIPFILES)): Remove unidiomatic for loop which frustrated
  parallel make operation.

Thanks to an anonymous contributor for the report and patch.

Fixes <https://savannah.gnu.org/bugs/?59958>.

The bug report characterizes the for loop as a "race", which I don't
think is strictly correct, since the outcome of the work done is not
subject to variation.  The body of the for loop is a clobbering shell
redirection operator, so the problem was one of potential pointless
repetition of work in concurrent operation.  Good to fix, but not a
cause for alarm.

Problem dates back to 2609cfb727a98809548edfd721c5cb923085b24b, 6
September 2014.