groff_ms(7): Reformat tables.
Instead of contriving an ad hoc sub-subsection heading with bold
paragraph text and a relative inset for each table, center the same text
and span the table width with it.  This makes the inset unnecessary and
recovers precious horizontal space for the entries.

Also ditch .ne requests, which are discouraged in groff man page style,
which had fallen out of sync with the actual table lengths, and which
are not reliable with boxed tables in nroff mode.