[ms]: Document origin/portability of .BI.
The availability of a bold-italic style for nearly any font family,
taken for granted in the PostScript era, is a late development relative
to AT&T troff and ms.  The V7 Unix and 4.2BSD ms implementations lack
the macro altogether.

V10 Research Unix has it, with a terse comment: "only on 202"; I surmise
that this is a reference to the equipment memorialized in the 1980
Condon/Kernighan/Thompson paper "Experience with the Mergenthaler
Linotron 202 Phototypesetter, or, How We Spent Our Summer Vacation"
<https://www.cs.princeton.edu/~bwk/202/index.html>.