[ms]: Clarify when the date must be declared.
Generalize; if someone uses the arguments to DA or ND macros, they'll
need to call them well before the end of the first page if they also
redefine one of the header strings to include it.  There are sufficient
variables here involving use of the header/footer-setting macros, their
corresponding strings, and the .P1 macro that it's easier just to be
abstract and trust users to experiment and/or think for themselves.