[man pages]: Use sentence case in section heads.
...as discussed on the groff development list[1].  Thanks to Ingo
Schwarze, John Gardner, and Peter Schaffter for the discussion.

Also simplify some section headings.
* "The Language" -> Language
* "Perl Parts" -> "Perl regions"[2]
* "groff Macro Packages" -> "Macro packages" (the context is obvious)
* "Full packages" -> "Full-service packages" (also used in groff.texi)
* "Simple Examples" -> "Examples" (after Linux man-pages 5.07)

And drop the entire subsection of gperl(1) which seems to be saying that
the standard error stream behaves normally.  This is news?!

[1] https://lists.gnu.org/archive/html/groff/2020-01/msg00063.html

[2] The previous author of this man page shares with the Ada language
standardization committee an unrelenting preoccupation with the word
"part", which always puts me in mind of a Sol Star/Seth Bullock
colloquy.

"He looked pale to me."

"What if he was?  Let's say he was.  Will you shut up about it?  What is
my part and your part?  What part of my part is your part?  Is my foot
your knee?  What about your ear?  What the **** is that?"