gropdf(1): Fix style nit.
When combining quotation with italics (or bold), the punctuation should
not share the styling of the quoted material.  This page uses italics
for predefined literals[1] that can be passed to the 'pdf: pagename' and
'pdf: switchtopage' device escapes.  But the directional single quotes
themselves around these literals should not be in italics, nor should
the trailing, sentence-ending punctuation--in the "logical quotation
style" generally favored by by groff documentation.

[1] incidentally illustrating that the consensus around
    bold-for-literals and italics-for-metavariables is not universal