[man]: Deprecation-warn on use of .DT, .PD.
Add warnings for macros deprecated in groff_man(7), emitted if the
CHECKSTYLE register is greater than 1.

* tmac/an-old.tmac (an-deprecation-warn): New macro emits diagnostic.

  (DT, PD): Call the foregoing, re-using arguments.

Be advised that our groff(7) page will trigger the .PD warning like
gangbusters, because a previous author of that page used it with .TP to
simulate the tables in the front matter of CSTR #54.